Isolate system settings types
