+++ src/models/system-settings/raw-system-settings.ts
@@ -1,0 +1,15 @@
+/*************************************************************************
+ * Copyright 2020 Gravwell, Inc. All rights reserved.
+ * Contact: <legal@gravwell.io>
+ *
+ * This software may be modified and distributed under the terms of the
+ * MIT license. See the LICENSE file for details.
+ **************************************************************************/
+
+export interface RawSystemSettings {
+	DisableMapTileProxy: boolean;
+	DistributedWebservers: boolean;
+	MapTileUrl: string;
+	MaxFileSize: number;
+	MaxResourceSize: number;
+}
