+++ src/models/system-settings/system-settings.ts
@@ -1,0 +1,17 @@
+/*************************************************************************
+ * Copyright 2020 Gravwell, Inc. All rights reserved.
+ * Contact: <legal@gravwell.io>
+ *
+ * This software may be modified and distributed under the terms of the
+ * MIT license. See the LICENSE file for details.
+ **************************************************************************/
+
+export interface SystemSettings {
+	mapTileURL: string;
+	disableMapTileProxy: boolean;
+
+	webServerIsDistributed: boolean;
+
+	maxFileSize: number;
+	maxResourceSize: number;
+}
